Improved the Christmas event. Localized lots of functions because they were either not needed for the outside world or improve speed. Cleaned up.

+++ lua/eventsystem/sv_core.lua
@@ -3,73 +3,65 @@ AddCSLuaFile("cl_core.lua")
 include("sh_core.lua")
 
 util.AddNetworkString("eventsystem_announce")
-util.AddNetworkString("eventsystem_cleanup")
-util.AddNetworkString("eventsystem_startevent")
-util.AddNetworkString("eventsystem_endevent")
-util.AddNetworkString("eventsystem_remove")
+util.AddNetworkString("eventsystem_start")
+util.AddNetworkString("eventsystem_end")
 
-function eventsystem:Announce(recipients, message, duration)
-	net.Start("eventsystem_announce", recipients)
+function eventsystem:Announce(message, duration, recipients)
+	net.Start("eventsystem_announce")
 		net.WriteString(message)
 		net.WriteInt(duration, 16)
-	net.Send(recipients)
-end
 
-function eventsystem:AnnounceEveryone(message, duration)
-	self:Announce(player.GetAll(), message, duration)
-end
-
-function eventsystem:CleanupClient(player)
-	net.Start("eventsystem_cleanup")
-	net.Send(player)
-end
-
-function eventsystem:CleanupEveryone()
-	net.Start("eventsystem_cleanup")
-	net.Broadcast()
+	if recipients then
+		net.Send(recipients)
+	else
+		net.Broadcast()
+	end
 end
 
 --There's 2 types of events as of now: string and event. Should be self explanatory.
-function eventsystem:AddScheduledEvent(evtype, data, time)
+function eventsystem:AddScheduled(evtype, data, time)
 	if evtype == "string" and #data >= 32672 then
-		error("scheduled code string needs to be smaller than 32672 bytes", 2)
+		error("scheduled code string needs to be smaller than 32672 bytes")
 	end
 
-	if istable(time) then
+	local timetype = type(time)
+	if timetype == "table" then
 		if not time.year and not time.month and not time.day then
-			error("required members year, month and day for the time table given to schedule event didn't exist", 2)
+			error("required members year, month and day for the time table given to schedule event didn't exist")
 		end
 
 		time = os.time(time)
-	elseif not isnumber(time) then
-		error("can't schedule event because provided time is not a valid type (type was " .. type(time) .. ")", 2)
+	elseif timetype ~= "number" then
+		error("can't schedule event because provided time is not a valid type (type was " .. timetype .. ")")
 	end
 
 	sql.Query("INSERT INTO eventsystem_schedules (Type, Data, Time) VALUES (" .. SQLStr(evtype) .. ", " .. SQLStr(runstring) .. ", " .. time .. ")")
 	return tonumber(sql.Query("SELECT LAST_INSERT_ID()"))
 end
 
-function eventsystem:RemoveScheduledEvent(number)
-	sql.Query("DELETE FROM eventsystem_schedules WHERE Number = " .. number)
-end
-
-function eventsystem:RemoveScheduledEvents()
-	MsgN("[Event System] Removing all scheduled events.")
-	sql.Query("DELETE * FROM eventsystem_schedules")
+function eventsystem:RemoveScheduled(number)
+	if number then
+		sql.Query("DELETE FROM eventsystem_schedules WHERE Number = " .. number)
+	else
+		MsgN("[Event System] Removing all scheduled events.")
+		sql.Query("DELETE * FROM eventsystem_schedules")
+	end
 end
 
 timer.Create("eventsystem_SchedulesChecker", 1, 0, function()
 	local tbl = sql.Query("SELECT * FROM eventsystem_schedules WHERE Time <= strftime('%s', 'now')")
-	if not tbl then return end
+	if not tbl then
+		return
+	end
 
-	for _, v in pairs(tbl) do
-		if v.EventType == "event" then
-			self:StartEvent(v.Data)
+	for _, event in pairs(tbl) do
+		if event.EventType == "event" then
+			self:Start(event.Data)
 		else
-			RunStringEx(v.Data, "Event System Scheduled RunStringEx")
+			RunStringEx(event.Data, "Event System scheduled RunStringEx")
 		end
 
-		self:RemoveScheduledEvent(v.Number)
+		self:RemoveScheduled(event.Number)
 	end
 end)
